Improve DB storage, don't save the content text in the DB, just the embeddings. Also, truncate the content text to first 8000 characters only.

+++ chrome-extension/bridge/ai-bridge.js
@@ -244,7 +244,7 @@ Remember: You can only reference information from the provided browsing history 
     const sessionOptions = {
       type: options.type || 'tldr', // Fixed typo: 'tl;dr' -> 'tldr'
       format: options.format || 'plain-text',
-      length: options.length || 'short'
+      length: options.length || 'long'
     };
 
     // Add monitor for download progress if requested
